Put SMI into it's own MAC module
Remove type state and just have another type for owned SMI
